requirements.txt
Flask==3.1.0
numpy==2.2.5
Pillow==11.2.1
Requests==2.32.3
torch==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
